refactor review delete
